update price data type to double

+++ database/migrations/2023_02_09_010906_create_properties_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->enum('unitType', ['Studio','Loft', 'Bedroom Unit', 'Penthouse Unit', 'Bi-Level Unit']);
             $table->enum('unitStatus', ['Sold','Rented', 'For Sale', 'For Rent']);
             $table->float('floorArea')->nullable();
-            $table->float('unitPrice');
+            $table->double('unitPrice');
             $table->integer('numBed')->nullable();
             $table->integer('numBaths')->nullable();
             $table->string('imagePath')->nullable()->default(null);
